Improve install script
The Klipper service only needs to be restarted on initial install. On
updates done through Moonraker, the Moonraker service will restart the
Klipper service based on the update configuration.

This commit enhances the install script to only restart Klipper if it
detects that the extensions have not been installed yet.