Improve interpolation by computing the factor
Previous version had the change factor for each step in the transition
as an input to the command. This is not very user-friendly as it's hard
for users to figure out what the factor should be in order to get the
desired duration.

This change improves the module by replacing the FACTOR argument with a
DURATION argument that can be used to specify how long the transition
should take.

The step factor is then computed by the module itself based on the
input duration and a frame rate of 24 frames per second.