Fix timer registration for delayed template execution timer
The module uses a timer to delay the execution of templates in cases
when the GCode object is locked (due to it currently running commands).

The timer is registered with the reactor if the template cannot be
executed right away. The timer fire time was set to be the current event
time plus a preset delay. However, due to the preset delay being very
small, it could result in the timer fire time being in the past (since
code execution takes time). This could result in the template never
getting executed.

Rather than using the event time to register the timer, use the current
monotonic time to compute the timer fire time.