Remove requirement for [menu] or [display] sections
State changes based on LCD menu button events requires that
the module registers handlers for menu events. However, on
printers that do not have LCD display, the configuration
will not include a `[menu]` and/or `[display]` sections.

This led to a Klipper crash when the module was initialized
since it expected to be able to lookup the `[menu]` object.

This commit fixes issue #4 reported for the above crash by
relaxing the requirement. Instead, registration for menu
events only happens if there is a `[menu]` object.