Fix activity detection right after printer restart
This commit fixes a corner case of the activity detection that happens
only if the printer becomes active right after a restart.

The issue is that the printer's idle_timeout module transitions into the
"Printing" state as part of the printer's initialization (any GCode or
macro executed during startup). This results in the
"idle_timeout:printing" event being sent. However, we don't want to catch
that event as it will result in a false activity detection.

However, if the printer becomes active prior to the idle_timeout module
having had the chance to transition out of "Printing" and into "Idle"
state, a new "idle_timeout:printing" event is not sent.

As a result, the state_notify module can miss the detection of the
activity and erroneously trigger the "inactive" state even though the
printer is active.

To fix this, the state_notify module is now also registering for the
"toolhead:sync_print_time" event. This event handler will be called on
any activity, which allows us to catch activity right after restart.
This does end up generating a lot of events, so in order to avoid
polluting the log with unwanted prints, the log entry on sub-state
changes has been commented out.