Combine "idle_timeout" and "state_notify" idle GCode
The 'state_notify' module allows for GCode to be executed when the
printer enters the "idle" state. This GCode was in addition to the GCode
already being executed by the 'idle_timeout:gcode' configuration.

However, due to how 'state_notify' transitions to the "idle" state, the
'state_notify:on_idle_gcode' GCode gets executed after the 'idle_timeout'
module transitions to the "Idle" state. This causes the printer to jump
right back into the 'active' state.

To fix this, override the 'idle_timeout:gcode' GCode with the combined
GCode from that section and 'state_notify:on_idle_gcode'. With this
change, all "idle" GCode is executed as part of the 'idle_timeout:gcode'
template, prior to the transition to the "Idle" state.