Handle a corners case for a transition to "active" state
A corner case exists, which can happen if the pritner starts executing
commands right after a transition to "Ready" state. If this happens,
the 'state_notify' module misses the notification and erronously goes
into the "inactive" state even though the printer is active.

This happens due to the fact that the 'idle_timeout' state goes to
"Printing" because the execution of the 'state_notify:on_ready_gcode'
template. Then, the commands that are queued start being executed.

However, since the 'idle_timeout' module is already in the "Printing"
state from the "on_read_gcode" template, it just stays in "Printing"
state. Therefore, 'state_notify' never gets a notification that the
printer has transition from 'ready' to 'active'.

To fix this, register for the 'toolhead:sync_print_time' event but only
pay attention to it once. The event handler will transition the printer
to the active state but only on the first trigger of the event.

So, if the handler is called while the printer is in the 'ready' state,
it will be transition to 'active', closing the hole exposed by the
corner case. If the printer is already in the 'active' state, the event
is ignored.