Fix state change during toolhead sync evnet
The state that the toolhead sync event was using to close the corner
case described in commit 54d34cc ("Handle a corners case for a
transition to "active" state") was "active".

The issue with using that state is that it does not correctly cancel the
inactive timer, resulting in the 'state_notify' state switching to the
'inactive' state while the printer is still active.

Using the 'idle_priting' state mimics what the 'idle_timeout' module
would be using and fixes the issue.