[state_notify] Run the "active" template when a print is done
Run the "on_active_gcode" template when the printer transitions from
"printing" to "active" after a print is done.

Doing this requires that the inactive timer is reschedule after the
state transition. Otherwise, the substate switch to "idle_printing" (a
result of the template execution) will cause the timer to be cancelled
right after having been rescheduled.