[loop_macro] Add support for iteration limits
Loop macros have the potential of running forever and locking up
Klipper completely. This commit attemtps to improve this by providing
a way for loop macros to define a safety switch by limiting the number
of iterations the macro can go through.

While it is still possible to lock up Klipper if limits are not defined,
if a limit is in place, a loop macro will terminate even if it does not
reach a `BREAK` command.