[settling_probe] Update settling_probe extension after Klipper probe changes
Klipper has recently completely re-organized the probe.py module. This
broke the settling_probe extensions since it reaches in the probe objects.
Due to the changes, the internal manupulations that settling_probe was
doing did not work and would cause Klipper crashes.

This change fixes this by reworking the settling_probe extension to work
with the newly re-organized probe code.