[state_notify] Detect print resume
The module successfully detects when a printer is paused and changes
the state accordingly. However, it did not properly detect when the
print has resumed.

As a result, the state remained "paused" until the print ended and the
state was changed to "active".

This change fixes this by monitoring the print statistics, which more
accurately describe the current state of the print.