[state_notify] Use `klippy:mcu_identify` to register `klippy:ready` handler
Klipper executes the `klippy:ready` handlers in the order in which they
have been registered. What this means is that any objects, which are
defined after the `[state_notify]` section and register `klippy:ready`
handlers will have their handlers execute after `state_notify`'s
handler.

This could result in a Klipper crash if the `state_notify` `klippy:ready`
handler executes GCode that references any of those objects.

This change attempts to handler this by changing how the `state_notify`
handler is registered. Instead of registering it in the object
constructor, the module waits for the `klippy:mcu_identify` event and
registers the `klippy:ready` event there.