[led_interpolate] Update extension to latest Klipper
Some recent changes to Klipper reworked how the LEDHelper object works.
It is no longer a container for all LED objects of different types.
Rather it is a pure helper that LED objects use. This prevents use from
using it to find the LEDs specified in the interpolation command.

Of course, the real cause of this is that Klipper does not have a
plugin system so it has no stable ABI.

This change updates the led_interpolate module to work with the current
state of LED support. It is a fairly big rework, which also simplifies
the internal state and the interpolation algorithm.