[state_notify] Handling missing `idle_tieout` configuration
The extension states that the `idle_timeout` module is automatically
loaded by the extension. This implies that there is no need for any
`idle_timeout` configuration to exist in the config files.

However, without any `idle_timeout` configuration, the extension causes
a Klipper crash because it tries to look up the configuration, which
does not exist.

The above was reported in bug #14. Fix the bug by actually loading the
`idle_timeout` module and safely looking up the module's configuration.