[state_notify] Add `heaters_active` option
Add a new option to the `state_notify` module that would keep the
printer in the active state if there are any active heaters.

The purpose for this option is to allow printer preheating,
especially chamber preheating. Normally, getting the printer
chamber upto temp takes a long time. If a user wants to preheat
the chamber before starting any prints, setting the bed temp and
letting the printer start heating might be desirable.

Unfortunately, since heater activity is not the same as toolhead
activity, it does not affect Klipper's idle time. This will
result in Klipper turning off the heaters when the timeout
expires.

Of course, this could be handled by using the `SET_IDLE_TIMEOUT`
command to extend the timeout or simply makeing the default
timeout longer. However, I find using any of those two methods
suboptimal.

Note that using this option would cause heaters to be active
until they are explicitly turned off.