[state_notify] Correctly handle initial transition from "Ready"
Commit 54d34cc ("Handle a corners case for a transition to "active"
state") added a toolhead sync handler to handle a supposed corner case
that misses a transition to `active` state on initial printer startup.

This not only caused some issue with proper state transitions but was
also completely unnecessary.

With this handler in place, Klipper would issue the toolhead sync event
when the printer executes any template. If the idle timeout is shorted
than the `inactive` timeout, Klipper's `idle_timeout` Gcode execution
would cause the toolhead sync even to trigger. This, in turn, would put
the printer back in to the `active` state rather than the `idle` state.

What's even more "wrong" about this event handler is that it was
completely unnecessary. Instead of using it to transition the printer
into the active state based on an intial toolhead sync event, it makes
more sense to transition it right after the `ready` state:

  * The `klippy:ready` handler would transition the state to `ready` and
    execute the `on_ready_gcode` template.
  * After executing the `on_read_gcode` template, the state is
    automatically transitioned to the `ready` state and the
    `on_active_gcode` template is executed.