[state_notify] Keep printer in `active` state while menu is active
While test with short `idle_timeout:timeout` values, it turned out that
a running menu did not keep the printer in `active` state. The
`idle_timeout:idle` state triggered and put the printer in the `idle`
state.

To fix this, the extension needs to keep "trick" the `idle_timeout`
module into thinking that the printer is still active. This can be done
by running a `G4` command, which doesn't do anything. However, it does
update the toolhead state, which causes the `idle_timeout` module to
consider the printer active.

So, now the timer that check if the menu is still running would also
execute this "noop" template if the menu is still running.